backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
groq
google-genai
edge-tts
pydub
